+++ app/src/at-hungtran/java/asiantech/internship/summer/recyclerview/PagerActivity.java
@@ -3,9 +3,6 @@ package asiantech.internship.summer.recyclerview;
 import android.annotation.SuppressLint;
 import android.os.Bundle;
 import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
@@ -28,42 +25,7 @@ public class PagerActivity extends AppCompatActivity{
             setSupportActionBar(toolbar);
         }
 
-        viewPager.setAdapter(new SectionPagerAdapter(getSupportFragmentManager()));
+        viewPager.setAdapter(new PagerAdapter(getSupportFragmentManager()));
         tabLayout.setupWithViewPager(viewPager);
     }
-
-    public class SectionPagerAdapter extends FragmentPagerAdapter {
-
-        SectionPagerAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            switch (position) {
-                case 0:
-                    return new TimelineFragment();
-                case 1:
-                default:
-                    return new FavouriteFragment();
-            }
-        }
-
-        @Override
-        public int getCount() {
-            return 2;
-        }
-
-        @Override
-        public CharSequence getPageTitle(int position) {
-            switch (position) {
-                case 0:
-                    return "Timeline";
-                case 1:
-                default:
-                    return "Favourite";
-            }
-        }
-    }
-
 }
