+++ app/src/at-hungtran/java/asiantech/internship/summer/recyclerview/PagerAdapter.java
@@ -1,4 +1,37 @@
 package asiantech.internship.summer.recyclerview;
 
-public class PagerAdapter {
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+public class PagerAdapter extends FragmentPagerAdapter {
+    public PagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        switch (position) {
+            case 0:
+                return new TimelineFragment();
+            case 1:
+            default:
+                return new FavouriteFragment();
+        }
+    }
+
+    @Override
+    public int getCount() {
+        return 2;
+    }
+    @Override
+    public CharSequence getPageTitle(int position) {
+        switch (position) {
+            case 0:
+                return "Timeline";
+            case 1:
+            default:
+                return "Favourite";
+        }
+    }
 }
